+++ Editor/CardEngineMenu/CreateCardSOWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 33088b0aff9b38f41ab60412cc3fc43f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
